more namespace refactoring
