+++ native-android/app/src/main/java/ai/protocol/ipfs/filforge/HomeFragment.kt
@@ -1,17 +1,11 @@
-package ai.protocol.ipfs.java
+package ai.protocol.ipfs.filforge
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.EditText
-import android.widget.Toast
 import androidx.compose.material3.Text
 import androidx.fragment.app.Fragment
-import io.ipfs.cid.Cid
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.ComposeView
