+++ native-android/app/src/main/java/ai/protocol/ipfs/filforge/MainActivity.kt
@@ -1,4 +1,4 @@
-package ai.protocol.ipfs.java
+package ai.protocol.ipfs.filforge
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
