+++ native-android/app/src/main/java/ai/protocol/ipfs/filforge/theme/DurinTheme.kt
@@ -1,6 +1,5 @@
-package ai.protocol.ipfs.java.theme
+package ai.protocol.ipfs.filforge.theme
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
