+++ native-android/app/src/test/java/ai/protocol/ipfs/filforge/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package ai.protocol.ipfs.java;
+package ai.protocol.ipfs.filforge;
 
 import org.junit.Test;
